+++ src/components/LabeledCheckBox/LabeledCheckBox.stories.tsx
